backend/resume_parser/requirements.txt
PyPDF2>=3.0.0
python-docx>=0.8.11
typing-extensions>=4.5.0 